// src/enums/data-test-id.ts
export enum DataTestId {
  card = 'card',
  loader = 'loader',
  error = 'error',
  navigationShowcase = 'navigation-showcase',
  navigationTerms = 'navigation-terms',
  navigationContract = 'navigation-contract',
  navigationBooks = 'navigation-books',
  burgerShowcase = 'burger-showcase',
  burgerTerms = 'burger-terms',
  burgerContract = 'burger-contract',
  burgerBooks = 'burger-books',
  burgerNavigation = 'burger-navigation',
  buttonHideReviews = 'button-hide-reviews',
  buttonRating = 'button-rating',
  buttonBurger = 'button-burger',
  buttonMenuViewWindow = 'button-menu-view-window',
  buttonMenuViewList = 'button-menu-view-list',
  buttonSearchOpen = 'button-search-open',
  buttonSearchClose = 'button-search-close',
  slideBig = 'slide-big',
  slideMini = 'slide-mini',
  inputSearch = 'input-search',
  sortRatingButton = 'sort-rating-button',
  searchResultNotFound = 'search-result-not-found',
  emptyCategory = 'empty-category',
  breadcrumbsLink = 'breadcrumbs-link',
  bookName = 'book-name',
  bookTitle = 'book-title',
  highlightMatches = 'highlight-matches',
  auth = 'auth',
  authForm = 'auth-form',
  registerForm = 'register-form',
  sendEmailForm = 'send-email-form',
  resetPasswordForm = 'reset-password-form',
  hint = 'hint',
  eyeClosed = 'eye-closed',
  eyeOpened = 'eye-opened',
  checkMark = 'checkmark',
  exitButton = 'exit-button',
  statusBlock = 'status-block',
  buttonRateBook = 'button-rate-book',
  buttonComment = 'button-comment',
  modalRateBook = 'modal-rate-book',
  bookingButton = 'booking-button',
  modalOuter = 'modal-outer',
  bookingModal = 'booking-modal',
  modalCloseButton = 'modal-close-button',
  modalTitle = 'modal-title',
  buttonPrevMonth = 'button-prev-month',
  buttonNextMonth = 'button-next-month',
  bookingCancelButton = 'booking-cancel-button',
  dayButton = 'day-button',
  calendar = 'calendar',
  monthSelect = 'month-select',
  rating = 'rating',
  star = 'star',
  starActive = 'star-active',
  alertClose = 'alert-close',
  editButton = 'edit-button',
  saveButton = 'save-button',
  historyReviewButton = 'history-review-button',
  cancelBookingButton = 'cancel-booking-button',
  emptyBlueCard = 'empty-blue-card',
  expired = 'expired',
  profilePage = 'profile-page',
  history = 'history',
  historySlide = 'history-slide',
  profileForm = 'profile-form',
  profileAvatar = 'profile-avatar',
  profileButton = 'profile-button',
  booking = 'booking',
  delivery = 'delivery',
  content = 'content',
  mainPage = 'main-page',
  comment = 'comment',
  commentAuthor = 'comment-author',
  commentWrapper = 'comment-wrapper',
  commentDate = 'comment-date',
  commentText = 'comment-text',
}